templates/.github/scripts/requirements.txt
requests==2.31.0
PyGithub==1.59.1
openai==1.3.8
transformers==4.35.2
huggingface_hub==0.17.3
groq==0.4.1
python-dotenv==1.0.0
pyyaml==6.0.1
gitpython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.4
tree-sitter-typescript==0.20.2
langchain==0.0.335
langchain-community==0.0.6
tiktoken==0.5.1 